+++ app/src/main/java/eu/kanade/domain/AnimeDomainModule.kt
@@ -12,9 +12,16 @@ import eu.kanade.domain.episode.interactor.SetExcludedAnimeScanlators
 import mihon.domain.animemigration.usecases.MigrateAnimeUseCase
 import tachiyomi.data.entries.anime.CustomAnimeRepositoryImpl
 import tachiyomi.data.entries.anime.AnimeRepositoryImpl
+import tachiyomi.data.entries.anime.AnimeMergeRepositoryImpl
 import tachiyomi.data.category.AnimeCategoryRepositoryImpl
 import tachiyomi.data.episode.EpisodeRepositoryImpl
+import tachiyomi.domain.entries.anime.interactor.DeleteAnimeById
+import tachiyomi.domain.entries.anime.interactor.DeleteByMergeId
+import tachiyomi.domain.entries.anime.interactor.DeleteMergeById
+import tachiyomi.domain.entries.anime.interactor.GetAllAnime
 import tachiyomi.domain.entries.anime.interactor.GetAnime
+import tachiyomi.domain.entries.anime.interactor.GetAnimeBySource
+import tachiyomi.domain.entries.anime.interactor.GetAnimeByUrlAndSourceId
 import tachiyomi.domain.entries.anime.interactor.FetchInterval
 import tachiyomi.domain.entries.anime.interactor.GetAnimeSeasonsByParentId
 import tachiyomi.domain.entries.anime.interactor.GetAnimeWithEpisodes
@@ -23,11 +30,19 @@ import tachiyomi.domain.entries.anime.interactor.GetDuplicateLibraryAnime
 import tachiyomi.domain.entries.anime.interactor.GetFavoriteAnime
 import tachiyomi.domain.entries.anime.interactor.GetFavorites
 import tachiyomi.domain.entries.anime.interactor.GetLibraryAnime
+import tachiyomi.domain.entries.anime.interactor.GetMergedAnime
+import tachiyomi.domain.entries.anime.interactor.GetMergedAnimeById
+import tachiyomi.domain.entries.anime.interactor.GetMergedAnimeForDownloading
+import tachiyomi.domain.entries.anime.interactor.GetMergedReferencesById
+import tachiyomi.domain.entries.anime.interactor.GetSeenAnimeNotInLibraryView
 import tachiyomi.domain.entries.anime.interactor.NetworkToLocalAnime
+import tachiyomi.domain.entries.anime.interactor.ResetViewerFlags
 import tachiyomi.domain.entries.anime.interactor.SetCustomAnimeInfo
 import tachiyomi.domain.entries.anime.interactor.SetAnimeEpisodeFlags
 import tachiyomi.domain.entries.anime.interactor.SetAnimeSeasonFlags
 import tachiyomi.domain.entries.anime.interactor.UpdateAnime as DomainUpdateAnime
+import tachiyomi.domain.entries.anime.interactor.UpdateMergedSettings
+import tachiyomi.domain.entries.anime.repository.AnimeMergeRepository
 import tachiyomi.domain.entries.anime.repository.AnimeRepository
 import tachiyomi.domain.entries.anime.repository.CustomAnimeRepository
 import tachiyomi.domain.source.anime.interactor.GetAnimeSourcesWithNonLibraryAnime
@@ -45,6 +60,7 @@ import eu.kanade.domain.episode.interactor.SyncEpisodesWithSource
 import tachiyomi.domain.episode.interactor.FilterEpisodesForDownload
 import tachiyomi.domain.episode.interactor.GetEpisode
 import tachiyomi.domain.episode.interactor.GetEpisodesByAnimeId
+import tachiyomi.domain.episode.interactor.GetMergedEpisodesByAnimeId
 import tachiyomi.domain.episode.interactor.SetAnimeDefaultEpisodeFlags
 import tachiyomi.domain.episode.interactor.SetSeenStatus as DomainSetSeenStatus
 import tachiyomi.domain.season.interactor.SetAnimeDefaultSeasonFlags
@@ -71,6 +87,9 @@ class AnimeDomainModule : InjektModule {
         addSingletonFactory<AnimeRepository> { AnimeRepositoryImpl(get()) }
         addSingletonFactory<CustomAnimeRepository> { CustomAnimeRepositoryImpl(get()) }
         addFactory { GetAnime(get()) }
+        addFactory { GetAllAnime(get()) }
+        addFactory { GetAnimeBySource(get()) }
+        addFactory { GetAnimeByUrlAndSourceId(get()) }
         addFactory { FetchInterval(get()) }
         addFactory { GetAnimeSeasonsByParentId(get()) }
         addFactory { GetAnimeWithEpisodes(get(), get()) }
@@ -85,11 +104,24 @@ class AnimeDomainModule : InjektModule {
         addFactory { SetAnimeEpisodeFlags(get()) }
         addFactory { SetAnimeSeasonFlags(get()) }
         addFactory { GetDuplicateLibraryAnime(get()) }
+        addFactory { DeleteAnimeById(get()) }
+        addFactory { ResetViewerFlags(get()) }
+        addFactory { GetSeenAnimeNotInLibraryView(get()) }
         addFactory { GetAnimeSourcesWithNonLibraryAnime(get()) }
+
+        addSingletonFactory<AnimeMergeRepository> { AnimeMergeRepositoryImpl(get()) }
+        addFactory { GetMergedAnime(get()) }
+        addFactory { GetMergedAnimeById(get()) }
+        addFactory { GetMergedReferencesById(get()) }
+        addFactory { UpdateMergedSettings(get()) }
+        addFactory { DeleteByMergeId(get()) }
+        addFactory { DeleteMergeById(get()) }
+        addFactory { GetMergedAnimeForDownloading(get()) }
 
         addSingletonFactory<EpisodeRepository> { EpisodeRepositoryImpl(get()) }
         addFactory { GetEpisode(get()) }
         addFactory { GetEpisodesByAnimeId(get()) }
+        addFactory { GetMergedEpisodesByAnimeId(get(), get()) }
         addFactory { GetAvailableAnimeScanlators(get()) }
         addFactory { GetExcludedAnimeScanlators(get()) }
         addFactory { SetExcludedAnimeScanlators(get()) }